[SPARK-36546][SQL] Add array support to union by name
This PR adds array of struct support to unionByName. It further tries to simplify some of the logic for re-projecting different types by creating a `mergeFields` method that can find the right method to merge various combinations of fields, currently structs or arrays, but could add maps in the future.

Currently unionByName doesn't support arrays of structs or maps of structs. This adds support for the arrays of structs and should make it easy to add maps of structs support in the future.

New capability to unionByName with arrays of structs.

New unit tests

Closes #34246 from Kimahriman/union-by-name-array.

Authored-by: Adam Binford <adamq43@gmail.com>
Signed-off-by: Liang-Chi Hsieh <viirya@gmail.com>